refactor(server): Phase 4 reorg — move lgnodes to nodes/* and rename to *.node.ts; fix imports/tests; clean merge artifacts (#194)
