+++ .changeset/phase-4-reorg.md
@@ -1,0 +1,5 @@
+---
+'apps-server': minor
+---
+
+refactor(server): move lgnodes to nodes/* and rename to *.node.ts; update imports and tests accordingly.
