+++ apps/server/__tests__/memory.connector.callmodel.wiring.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi } from 'vitest';
-import { CallModelNode } from '../src/lgnodes/callModel.lgnode';
+import { CallModelNode } from '../src/nodes/callModel.node';
 import { BaseTool } from '../src/tools/base.tool';
 import { AIMessage, BaseMessage, SystemMessage } from '@langchain/core/messages';
