+++ apps/server/src/agents/agent.ts
@@ -7,15 +7,15 @@ import { last } from 'lodash-es';
 import { McpServer, McpTool } from '../mcp';
 import { isDynamicConfigurable, type StaticConfigurable } from '../graph/capabilities';
 import { inferArgsSchema } from '../mcp/jsonSchemaToZod';
-import { CallModelNode, type MemoryConnector } from '../lgnodes/callModel.lgnode';
-import { ToolsNode } from '../lgnodes/tools.lgnode';
+import { CallModelNode, type MemoryConnector } from '../nodes/callModel.node';
+import { ToolsNode } from '../nodes/tools.node';
 import { CheckpointerService } from '../services/checkpointer.service';
 import { ConfigService } from '../services/config.service';
 import { LoggerService } from '../services/logger.service';
 import { BaseTool } from '../tools/base.tool';
 import { LangChainToolAdapter } from '../tools/langchainTool.adapter';
-import { SummarizationNode } from '../lgnodes/summarization.lgnode';
-import { EnforceRestrictionNode } from '../lgnodes/enforceRestriction.lgnode';
+import { SummarizationNode } from '../nodes/summarization.node';
+import { EnforceRestrictionNode } from '../nodes/enforceRestriction.node';
 import { NodeOutput } from '../types';
 import { z } from 'zod';
 import { stringify as toYaml } from 'yaml';
