+++ apps/server/src/agents/simple.agent.ts
@@ -7,18 +7,18 @@ import { last } from 'lodash-es';
 import { McpServer, McpTool } from '../mcp';
 import { isDynamicConfigurable } from '../graph/capabilities';
 import { inferArgsSchema } from '../mcp/jsonSchemaToZod';
-import { CallModelNode, type MemoryConnector } from '../lgnodes/callModel.lgnode';
-import { ToolsNode } from '../lgnodes/tools.lgnode';
+import { CallModelNode, type MemoryConnector } from '../nodes/callModel.node';
+import { ToolsNode } from '../nodes/tools.node';
 import { CheckpointerService } from '../services/checkpointer.service';
 import { ConfigService } from '../services/config.service';
 import { LoggerService } from '../services/logger.service';
 import { BaseAgent } from './base.agent';
 import { BaseTool } from '../tools/base.tool';
 import { LangChainToolAdapter } from '../tools/langchainTool.adapter';
-import { SummarizationNode } from '../lgnodes/summarization.lgnode';
+import { SummarizationNode } from '../nodes/summarization.node';
 import { NodeOutput } from '../types';
 import { z } from 'zod';
-import { EnforceRestrictionNode } from '../lgnodes/enforceRestriction.lgnode';
+import { EnforceRestrictionNode } from '../nodes/enforceRestriction.node';
 import { stringify as toYaml } from 'yaml';
 import { buildMcpToolError } from '../mcp/errorUtils';
