+++ apps/server/src/nodes/lgnodes/enforceRestriction.lgnode.ts
